"refactor: move modal to modal folder"
