+++ src/app/(ui)/barangay/[barangayName]/recipients/page.tsx
@@ -3,7 +3,7 @@
 import React, { useState } from "react";
 import { useRouter, useParams } from "next/navigation";
 import Table from "@/app/components/barangay/Table";
-import BeneficiaryModal from "@/app/components/barangay/BeneficiaryForm";
+import BeneficiaryModal from "@/app/components/barangay/modal/BeneficiaryForm";
 import { BeneficiaryForm } from "@/app/lib/definitions";
 
 // Define the type for the params object returned by useParams
